[ADD] command endpoint
