+++ .gitignore
@@ -25,3 +25,4 @@ go.work.sum
 .env
 .idea
 .internal/services/
+.test
