+++ README.md
@@ -1,4 +1,4 @@
-# gorock
+# GOROCK
 
 🌍 *[English](README.md) ∙ [Russian](README_ru.md)*<br>
 <img align="right" width="180px" src="media/icon.png">
@@ -39,11 +39,11 @@ export PATH=$GOBIN:$PATH
 ```sh
 go install github.com/arzab/gorock/cmd/gorock@latest
 ```
-3. Try to create you own `repository` fiberEndpoint by using
+3. Try to create you own `repository` service by using
 ```sh
-gorock fiberEndpoint repository
+gorock service repository
 ```
-4. Now you can use your fiberEndpoint in you project
+4. Now you can use your `service` in you project in any place (singleton template)
 ```go
 package main
