+++ internal/commands/cmd_delivery/cmd_endpoint/templates.go
@@ -1,0 +1,1 @@
+package cmd_endpoint
