+++ template/delivery/endpoints/endpoints.go
@@ -1,0 +1,14 @@
+package endpoints
+
+import (
+	"github.com/arzab/gorock/pkg/http/endpoints"
+)
+
+func HttpEndpoints() []endpoints.FiberEndpoint {
+	result := make([]endpoints.FiberEndpoint, 0)
+
+	//Endpoints
+	//Endpoints
+
+	return result
+}
