+++ template/services/configs.go
@@ -1,0 +1,6 @@
+package services
+
+type Configs struct {
+	// Configs
+	// Configs
+}
