+++ template/services/example/configs.go
@@ -1,0 +1,4 @@
+package example
+
+type Configs struct {
+}
