+++ template/services/example/interface.go
@@ -1,0 +1,4 @@
+package example
+
+type Service interface {
+}
