+++ template/services/example/service.go
@@ -1,0 +1,11 @@
+package example
+
+type service struct {
+	configs Configs
+}
+
+func NewService(configs Configs) Service {
+	return &service{
+		configs: configs,
+	}
+}
