+++ template/services/services.go
@@ -1,0 +1,9 @@
+package services
+
+var (
+// Var
+// Var
+)
+
+// Functions
+// Functions
